+++ flogin/default_events.py
@@ -27,6 +27,8 @@ async def on_error(
 def get_default_events(plugin: Plugin[Any]) -> dict[str, Callable[..., Awaitable[Any]]]:
     def on_query(data: RawQuery, raw_settings: dict[str, Any]):
         query = Query(data, plugin)
+        plugin.last_query = query
+
         if plugin._settings_are_populated is False:
             LOG.info(f"Settings have not been populated yet, creating a new instance")
             plugin._settings_are_populated = True